requirements.txt
# SDG Progress Tracker - Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Orchestration
mage-ai>=0.9.0

# Google Cloud / BigQuery
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0
db-dtypes>=1.1.1

# dbt
dbt-core>=1.7.0
dbt-bigquery>=1.7.0

# Machine Learning
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
